pyproject.toml
[tool.poetry]
name = "wishlist"
version = "0.1.0"
description = ""
authors = ["timosii <tsimashev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
aiogram = "^3.3.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
postgres = "^4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
